+++ src/db/repositories/list-repository.ts
@@ -4,6 +4,7 @@ import { schema } from '../schema'
 import type { GetListsInput } from '@/domain/services/lists/get-lists'
 import type { InsertListModel } from '@/domain/entities/lists'
 import type { UpdateListValues } from '@/domain/services/lists/update-list'
+import type { UpdateListBannerInput } from '@/domain/services/lists/update-list-banner'
 
 export function selectLists({
   userId,
@@ -90,10 +91,17 @@ export async function updateList(
 }
 
 export async function getListById(id: string) {
+  return db.select().from(schema.lists).where(eq(schema.lists.id, id))
+}
+
+export async function updateListBanner({
+  listId,
+  userId,
+  bannerPath,
+}: UpdateListBannerInput) {
   return db
-    .select({
-      ...getTableColumns(schema.lists),
-    })
-    .from(schema.lists)
-    .where(eq(schema.lists.id, id))
+    .update(schema.lists)
+    .set({ bannerPath })
+    .where(and(eq(schema.lists.id, listId), eq(schema.lists.userId, userId)))
+    .returning()
 }