+++ CHANGELOG.md
@@ -8,6 +8,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 
 ### Added
 
+- Create list items create, read, update and delete endpoints [#26](https://github.com/plotwist-app/plotwist-backend/pull/26)
 - Create get list endpoint [#22](https://github.com/plotwist-app/plotwist-backend/pull/22)
 - Create edit list endpoint [#21](https://github.com/plotwist-app/plotwist-backend/pull/21)
 - Create delete list endpoint [#20](https://github.com/plotwist-app/plotwist-backend/pull/20)